+++ features/recharge/digital_topup_bills/src/main/java/com/tokopedia/topupbills/telco/view/fragment/DigitalTelcoPrepaidFragment.kt
@@ -91,7 +91,7 @@ class DigitalTelcoPrepaidFragment : DigitalBaseTelcoFragment() {
                 }
 
                 productId = it.product.id.toIntOrZero()
-                price = it.product.attributes.pricePlain.toLong()
+                price = it.product.attributes.pricePlain
                 checkVoucherWithDelay()
 
                 checkoutPassData = DigitalCheckoutPassData.Builder()
